ENH: Tab removal and add final patient information Queries
Add queries for the immunizations and the CPT Code files that are
similarly filtered by a patient IEN.

Add a new option in the dropdown menu to ensure that at least one option
has information in these two new queries.

Add the files to index.html to handle and process the new returns.

Find and replace all \t in the two files and replace them with two
spaces.