ENH: Add a new date node to connect information
Add a new node to the graph which is named by a particular date/time.
This node should be used to connect entries across the data types that
have the same date/time in them.  The date/time string to be compared
against should be the date/time of a visit.

IE, A visit and the corresponding CPT code or a visit and the
corresponding immunization.